// taxe-calculator-app/src/app/enums/taxValue-enum.ts
export enum taxValues{
    tiers1 = 18200,
    tiers2 = 37000,
    tiers3 = 87000,
    tiers4 = 180000
}

export enum fixedAmount{
    tiers0 = 0,
    tiers1 = 3272,
    tiers2 = 19822,
    tiers3 = 54232
}

export enum percentage{
    tiers1 = 19,
    tiers2 = 32.5,
    tiers3 = 37,
    tiers4 = 45
}